Update hand-tracking module to BayesFilters 0.9.100.

+++ src/hand-tracking/include/MeshModel.h
@@ -20,7 +20,7 @@ public:
 
     virtual std::tuple<bool, std::string> getShaderPaths() = 0;
 
-    virtual std::tuple<bool, std::vector<Superimpose::ModelPoseContainer>> getModelPose(const Eigen::Ref<const Eigen::MatrixXf>& cur_states) = 0;
+    virtual std::tuple<bool, std::vector<Superimpose::ModelPoseContainer>> getModelPose(const Eigen::Ref<const Eigen::MatrixXd>& cur_states) = 0;
 };
 
 #endif /* MESHMODEL_H */
