+++ src/hand-tracking/include/NormTwoChiSquare.h
@@ -13,7 +13,7 @@ public:
 
     ~NormTwoChiSquare() noexcept;
 
-    std::pair<bool, Eigen::VectorXf> likelihood(const bfl::MeasurementModel& measurement_model, const Eigen::Ref<const Eigen::MatrixXf>& pred_states) override;
+    std::pair<bool, Eigen::VectorXd> likelihood(const bfl::MeasurementModel& measurement_model, const Eigen::Ref<const Eigen::MatrixXd>& pred_states) override;
 
 private:
     struct ImplData;
