+++ src/hand-tracking/src/cpu/NormTwoKLD.cpp
@@ -33,7 +33,7 @@ NormTwoKLD::~NormTwoKLD() noexcept
 { }
 
 
-std::pair<bool, VectorXf> NormTwoKLD::likelihood(const MeasurementModel& measurement_model, const Ref<const MatrixXf>& pred_states)
+std::pair<bool, VectorXd> NormTwoKLD::likelihood(const MeasurementModel& measurement_model, const Ref<const MatrixXd>& pred_states)
 {
     throw std::runtime_error("[NormTwoKLD][CPU] Unimplemented.");
 }
